Handle Reorgs in a separate function and remove logix from advanceFinalized
